api/requirements-ci.txt
# Optimized requirements for CI/CD - avoids grpcio compilation issues
# Azure Functions Core (worker provided by runtime)
azure-functions==1.18.0

# Azure Services
azure-cosmos==4.5.1
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-storage-blob==12.19.0

# Core Dependencies
pydantic==2.5.2
httpx==0.25.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0

# Testing Dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
requests>=2.31.0
python-dotenv>=1.0.0
pyjwt>=2.8.0

# Additional testing utilities
mock>=4.0.0
responses>=0.23.0

# Code quality tools
flake8>=6.1.0

# Note: azure-functions-worker is excluded to avoid grpcio compilation issues
# The Azure Functions runtime provides this automatically in production
